调整代码接口 保持与其它客户端 服务端一直 修改说明
